requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-vlc==3.0.21203
requests==2.31.0
python-dotenv==1.0.0
httpx==0.25.1


